fix: fixed create new inventory

+++ domain/inventory.model.go
@@ -80,12 +80,12 @@ type InventoryDetailParam struct {
 }
 
 type InventoryDetailStatus struct {
-	Scanned           int     `gorm:"scanned" json:"scanned"`
-	Shortage          int     `gorm:"shortage" json:"shortage"`
-	Surplus           int     `gorm:"surplus" json:"surplus"`
-	All               int     `gorm:"all" json:"all"`
-	New               int     `gorm:"new" json:"new"`
-	Accepted          int     `gorm:"accepted" json:"accepted"`
+	Scanned           float64 `gorm:"scanned" json:"scanned"`
+	Shortage          float64 `gorm:"shortage" json:"shortage"`
+	Surplus           float64 `gorm:"surplus" json:"surplus"`
+	All               float64 `gorm:"all" json:"all"`
+	New               float64 `gorm:"new" json:"new"`
+	Accepted          float64 `gorm:"accepted" json:"accepted"`
 	ShortageSupplySum float64 `gorm:"shortage_supply_sum" json:"shortage_supply_sum"`
 	ShortageRetailSum float64 `gorm:"shortage_retail_sum" json:"shortage_retail_sum"`
 	SurplusSupplySum  float64 `gorm:"surplus_supply_sum" json:"surplus_supply_sum"`
